cli/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "schist"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["python-frontmatter>=1.1.0", "pyyaml>=6.0"]

[project.scripts]
schist = "schist.__main__:main"
